core-99 Refactor code
