+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/BelphegorDisplayStrings.properties
@@ -26,6 +26,7 @@ betaTagRemovedSuccess = Beta tag removed successfully!
 appStats = App Stats
 communities=Communities
 appPublicationRequests = App publication Requests
+publicationRequestSuccess = {0} published successfully!
