+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/apps/AdminApps.gwt.xml
@@ -37,4 +37,9 @@
         <when-type-is class="org.iplantc.de.admin.apps.client.views.grid.cells.AdminAppNameCell.AdminAppNameCellAppearance" />
     </replace-with>
 
+    <replace-with class="org.iplantc.de.theme.base.client.admin.apps.AppPublicationRequestDefaultAppearance">
+        <when-type-is class="org.iplantc.de.admin.apps.client.AppPublicationRequestView.AppPublicationRequestAppearance" />
+    </replace-with>
+
+
 </module>
